- Fixing rename bug 2/2.
