+++ src/constants/CourseUrlSuffix.ts
